backend/requirements.txt
python-dotenv==1.0.1
requests==2.26.0
django==3.2
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
django-filter==2.4.0